requirements.txt
fastapi~=0.115.8
crypto~=1.4.1
SQLAlchemy~=2.0.38
boto3~=1.36.17
botocore~=1.36.17
python-dotenv~=1.0.1
PyJWT~=2.10.1
passlib~=1.7.4
pydantic~=2.10.6
pytest~=8.3.4
httpx~=0.28.1